[backend] refactor: 멤버 삭제(초대 거절) 로직 계층 분리 및 구조 개선 (Controller → Service → Repository → Mapper)

+++ backend/src/main/java/com/example/demo/member/MemberMapper.java
@@ -20,7 +20,7 @@ public interface MemberMapper {
 
     void updateMember(MemberUpdateRequestDTO member);  // 멤버 수정
 
-    void deleteMember(int memberId);    // 멤버 삭제
+    void deleteMember(long memberId);    // 멤버 삭제
 
     List<TeamMembersNameResponse> requestTeamMembersName(int teamId);  // 팀의 모든 멤버 이름 요청
